Added: 2 New Commands
Added the command wcs_setfx norecoilsm which as the name indicates is the SM version of the already existing Source-Python setfx norecoil command. The reason this new one was created is because the one in Source-Python seem to negatively impact the server's performance, causing massive server stress when paired with higher attackspeed races.

Also added the wcs_setfx automaticpistols, which allow the player to use any pistol as if it was an automatic weapon.